Segregate tests in different files.
