+++ pontoon/machinery/tests/conftest.py
@@ -12,14 +12,6 @@ def ms_api_key(settings):
 
 
 @pytest.fixture
-def ms_locale(locale_a):
-    """Set the Microsoft API locale code for locale_a"""
-    locale_a.ms_translator_code = 'gb'
-    locale_a.save()
-    return locale_a
-
-
-@pytest.fixture
 def google_translate_api_key(settings):
     """Set the settings.GOOGLE_TRANSLATE_API_KEY for this test"""
     key = '2fffff'
