+++ pontoon/test/fixtures/base.py
@@ -68,6 +68,14 @@ def google_translate_locale(locale_a):
 
 
 @pytest.fixture
+def ms_locale(locale_a):
+    """Set the Microsoft API locale code for locale_a"""
+    locale_a.ms_translator_code = 'gb'
+    locale_a.save()
+    return locale_a
+
+
+@pytest.fixture
 def locale_b():
     return factories.LocaleFactory(
         code="gs",
